Add translation from library objects to driver
Add Driver3Translator class which takes a library object and translates
it to Java Driver (3) representation.

Added tuple translation from driver to library. Added getTuple() to
AbstractField.

Fixed bug that resulted in wrong equals for AbstractField, Field and
Cell. Fixed Duration field types.

Changed replicator to use the translator instead of
TEMPORARY_PORTING_getAsDriverObject(). This temporary method is now
removed.

+++ scylla-cdc-base/src/main/java/com/scylladb/cdc/model/worker/RawChange.java
@@ -41,21 +41,21 @@ public interface RawChange {
      */
     class Duration {
         // months, days and nanoseconds
-        private final long months;
-        private final long days;
+        private final int months;
+        private final int days;
         private final long nanoseconds;
 
-        public Duration(long months, long days, long nanoseconds) {
+        public Duration(int months, int days, long nanoseconds) {
             this.months = months;
             this.days = days;
             this.nanoseconds = nanoseconds;
         }
 
-        public long getMonths() {
+        public int getMonths() {
             return months;
         }
 
-        public long getDays() {
+        public int getDays() {
             return days;
         }
